@@ -2,12 +2,11 @@
   "name": "todo-in-life",
   "version": "1.0.0",
   "description": "",
-  "main": "webpack.config.js",
+  "type": "module",
   "scripts": {
-    "start": "webpack serve --mode development",
-    "dev": "webpack --mode development",
-    "prod": "webpack --mode production",
-    "analyze": "webpack --mode production --analyze"
+    "dev": "node dev-server/server",
+    "build": "vite build",
+    "preview": "cross-env NODE_ENV=production node server"
   },
   "repository": {
     "type": "git",
@@ -20,18 +19,6 @@
     "url": "https://github.com/Akuu29/todo-in-life/issues"
   },
   "homepage": "https://github.com/Akuu29/todo-in-life#readme",
-  "devDependencies": {
-    "@types/react": "^18.0.9",
-    "@types/react-datepicker": "^4.4.2",
-    "@types/react-dom": "^18.0.5",
-    "css-loader": "^6.7.1",
-    "style-loader": "^3.3.1",
-    "ts-loader": "^9.3.0",
-    "typescript": "^4.6.4",
-    "webpack-bundle-analyzer": "^4.5.0",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.9.0"
-  },
   "dependencies": {
     "@emotion/react": "^11.9.0",
     "@fortawesome/fontawesome-svg-core": "^6.1.1",
@@ -41,10 +28,28 @@
     "emotion-reset": "^3.0.1",
     "react": "^18.1.0",
     "react-cookie": "^4.1.1",
-    "react-datepicker": "^4.8.0",
+    "react-datepicker": "^4.25.0",
     "react-dnd": "^16.0.1",
     "react-dnd-html5-backend": "^16.0.1",
-    "react-dom": "^18.1.0",
-    "webpack": "^5.72.1"
+    "react-dom": "^18.1.0"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.21",
+    "@types/node": "^20.9.0",
+    "@types/react": "^18.2.37",
+    "@types/react-datepicker": "^4.19.4",
+    "@types/react-dom": "^18.2.15",
+    "@typescript-eslint/eslint-plugin": "^6.10.0",
+    "@typescript-eslint/parser": "^6.10.0",
+    "@vitejs/plugin-react-swc": "^3.5.0",
+    "compression": "^1.7.4",
+    "cross-env": "^7.0.3",
+    "eslint": "^8.53.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.4",
+    "express": "^4.18.2",
+    "sirv": "^2.0.3",
+    "typescript": "^5.2.2",
+    "vite": "^5.0.0"
   }
 }
