@@ -1,5 +1,9 @@
-import ReactDOM from "react-dom";
+import React from "react";
+import ReactDOM from "react-dom/client";
 import LoginPage from "./views/Login/LoginPage";
 
-const login = document.getElementById("login");
-ReactDOM.render(<LoginPage />, login);
+ReactDOM.createRoot(document.getElementById("login")!).render(
+  <React.StrictMode>
+    <LoginPage />
+  </React.StrictMode>
+);
