@@ -1,8 +1,8 @@
-import { FC } from "react";
 import { css } from "@emotion/react";
 
 const footer = css({
   width: "100%",
+  paddingTop: 140,
 });
 
 const footerTitle = css({
@@ -11,12 +11,12 @@ const footerTitle = css({
   margin: 50,
 });
 
-const Footer: FC = () => {
+function Footer() {
   return (
     <div css={footer}>
       <h4 css={footerTitle}>TODO IN LIFE@2022</h4>
     </div>
   );
-};
+}
 
 export default Footer;
