@@ -1,4 +1,3 @@
-import { FC } from "react";
 import { CookiesProvider } from "react-cookie";
 import GlobalStyles from "../../components/layout/GlobalStyles/GlobalStyles";
 import Header from "../../components/layout/Header/Header";
@@ -6,7 +5,7 @@ import Footer from "../../components/layout/Footer/Footer";
 import UserFormErrorMessages from "../../components/forms/UserFormErrorMessages/UserFormErrorMessages";
 import LoginForm from "./LoginForm";
 
-const LoginPage: FC = () => {
+function LoginPage() {
   return (
     <div>
       {/* CookiesProviderでラップしているコンポーネント内でcookieの管理が可能*/}
@@ -19,6 +18,6 @@ const LoginPage: FC = () => {
       </CookiesProvider>
     </div>
   );
-};
+}
 
 export default LoginPage;
