@@ -1,4 +1,4 @@
-import { FC, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { css } from "@emotion/react";
 import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
@@ -8,6 +8,7 @@ import TodoDescription from "./TodosTodoDescription";
 import TodosForm from "./TodosForm";
 import { TodoApi } from "../../services/api/todoApi";
 import { TODO_CATEGORIES } from "../../utils/constants/todoCategory.constants";
+import { useTodo } from "../../components/context/TodoContext";
 
 const todosContainer = css({
   marginTop: 40,
@@ -16,66 +17,63 @@ const todosContainer = css({
   justifyContent: "space-around",
 });
 
-export interface Todo {
-  id: string;
-  title: string;
-  content: string;
-  category: string;
-  date_limit: string | null;
-  date_created: string;
-}
-
-export type Todos = CustomObject<Array<Todo>>;
-
-const TodosContens: FC = () => {
-  const [todo, setTodo] = useState<Todo>({
-    id: "",
-    title: "",
-    content: "",
-    category: "",
-    date_limit: null,
-    date_created: "",
-  });
-
-  const [todos, setTodos] = useState<Todos>({
-    [TODO_CATEGORIES.SHORT]: [],
-    [TODO_CATEGORIES.MEDIUM]: [],
-    [TODO_CATEGORIES.LONG]: [],
-    [TODO_CATEGORIES.COMPLETE]: [],
-  });
+function TodosContents() {
+  const { todosByCategory, setTodosByCategory } = useTodo();
 
   useEffect(() => {
-    // todoの一覧の設定
     const setInitialTodos = async () => {
-      let initialTodosData: Todos = {
-        [TODO_CATEGORIES.SHORT]: [],
-        [TODO_CATEGORIES.MEDIUM]: [],
-        [TODO_CATEGORIES.LONG]: [],
-        [TODO_CATEGORIES.COMPLETE]: [],
-      };
+      const response = await TodoApi.getTodos();
+      if (response) {
+        const body = response.data;
 
-      const getTodosResult = await TodoApi.getTodos();
-      if (getTodosResult) {
-        const data = getTodosResult.data;
-        if (data.status == "success") {
-          data.todos.forEach((todo: Todo) => {
-            const key: keyof Todos = todo.category;
-            initialTodosData[key].push(todo);
+        if (body.status == "success") {
+          const todos = body.todos;
+          setTodosByCategory({
+            [TODO_CATEGORIES.SHORT]: todos.filter(
+              (todo) => todo.category == TODO_CATEGORIES.SHORT
+            ),
+            [TODO_CATEGORIES.MEDIUM]: todos.filter(
+              (todo) => todo.category == TODO_CATEGORIES.MEDIUM
+            ),
+            [TODO_CATEGORIES.LONG]: todos.filter(
+              (todo) => todo.category == TODO_CATEGORIES.LONG
+            ),
+            [TODO_CATEGORIES.COMPLETE]: todos.filter(
+              (todo) => todo.category == TODO_CATEGORIES.COMPLETE
+            ),
           });
         } else {
           // AxiosError
-          alert(`ERROR: ${getTodosResult.status}`);
+          alert(`ERROR: ${response.status}`);
         }
       } else {
         // Error
         alert("ERROR");
       }
-
-      setTodos(initialTodosData);
     };
 
     setInitialTodos();
   }, []);
+
+  useEffect(() => {
+    const setMaxPage = () => {
+      Object.keys(todosByCategory).forEach((key) => {
+        const todosLen = todosByCategory[key].length;
+        const todosMaxPage = Math.ceil(todosLen / 6) - 1;
+        if (key == TODO_CATEGORIES.SHORT) {
+          setPageShort([currentPageShort, todosMaxPage]);
+        } else if (key == TODO_CATEGORIES.MEDIUM) {
+          setPageMedium([currentPageMedium, todosMaxPage]);
+        } else if (key == TODO_CATEGORIES.LONG) {
+          setPageLong([currentPageLong, todosMaxPage]);
+        } else {
+          setPageCompleted([currentPageCompleted, todosMaxPage]);
+        }
+      });
+    };
+
+    setMaxPage();
+  }, [todosByCategory]);
 
   // それぞれのカテゴリ内における現在のページと最大のページ数
   const [[currentPageShort, maxPageShort], setPageShort] = useState<
@@ -91,28 +89,6 @@ const TodosContens: FC = () => {
     [number, number]
   >([0, 0]);
 
-  // todosの更新時に実行される
-  useEffect(() => {
-    // 各カテゴリの最大ページの設定
-    const setMaxPage = () => {
-      Object.keys(todos).forEach((key) => {
-        let todosLen = todos[key].length;
-        let todosMaxPage = Math.ceil(todosLen / 6) - 1;
-        if (key == TODO_CATEGORIES.SHORT) {
-          setPageShort([currentPageShort, todosMaxPage]);
-        } else if (key == TODO_CATEGORIES.MEDIUM) {
-          setPageMedium([currentPageMedium, todosMaxPage]);
-        } else if (key == TODO_CATEGORIES.LONG) {
-          setPageLong([currentPageLong, todosMaxPage]);
-        } else {
-          setPageCompleted([currentPageCompleted, todosMaxPage]);
-        }
-      });
-    };
-
-    setMaxPage();
-  }, [todos]);
-
   // todo詳細ページの表示非表示制御
   const [isShowTodoDesc, setIsShowTodoDesc] = useState<boolean>(false);
 
@@ -126,25 +102,16 @@ const TodosContens: FC = () => {
   // editとしてフォームが開かれた場合に、edit前のtodoのcategoryを保持しておく
   const [prevTodoCategory, setPrevTodoCategory] = useState<string>("");
 
-  // 各'AppTodosCategoryColumn'にtodoを描画する
+  // TodosColumnにtodoを描画する
   const returnTodosForColumn = (columnName: string, page: number) => {
-    if (todos[columnName].length) {
-      // 1pageに6件表示
-      return todos[columnName]
-        .filter((_, i) => i >= 6 * page && i < 6 * (page + 1))
-        .map((todo) => (
+    if (todosByCategory[columnName].length) {
+      return todosByCategory[columnName]
+        .filter((_, i) => i >= 10 * page && i < 10 * (page + 1))
+        .map((todoByCategory) => (
           <TodoContents
-            key={todo.id}
-            id={todo.id}
-            title={todo.title}
-            content={todo.content}
-            category={todo.category}
-            date_limit={todo.date_limit}
-            date_created={todo.date_created}
-            todos={todos}
-            setTodos={setTodos}
+            key={todoByCategory.id}
+            todo={todoByCategory}
             setIsShowTodoDesc={setIsShowTodoDesc}
-            setTodo={setTodo}
             setPrevTodoCategory={setPrevTodoCategory}
           />
         ));
@@ -158,181 +125,23 @@ const TodosContens: FC = () => {
     date_limit: [],
   });
 
-  // todoの作成。作成後、画面に反映
-  const createTodo: FnToHandleTodosTable = async () => {
-    const postTodoResult = await TodoApi.postTodo(todo);
-
-    if (postTodoResult) {
-      const data = postTodoResult.data;
-      if (data.status == "success") {
-        const todoCreated = data.todo;
-        setTodos({
-          ...todos,
-          [todoCreated.category]: [...todos[todoCreated.category], todoCreated],
-        });
-
-        // フォーム画面を閉じる
-        setIsShowForm(false);
-      } else {
-        // AxiosError
-        if (data.validationErrors) {
-          // バリデーションエラー
-          const validationErrors = data.validationErrors;
-          handleValidationErrors(validationErrors);
-        } else {
-          alert(`ERROR: ${postTodoResult.status}`);
-        }
-      }
-    } else {
-      // Error
-      alert("ERROR");
-    }
-  };
-
-  // todoの編集。編集内容を画面に反映
-  const editTodo: FnToHandleTodosTable = async () => {
-    const putTodoResult = await TodoApi.putTodo(todo);
-
-    if (putTodoResult) {
-      const data = putTodoResult.data;
-      if (data.status == "success") {
-        const todoEdited = data.todoEdited;
-        setTodos((prevTodos) => {
-          // edit対象となるcategoryのtodo配列の取得
-          let targetTodoArray = prevTodos[prevTodoCategory];
-
-          if (todoEdited.category == prevTodoCategory) {
-            // edit前と後でcategoryが同じ場合
-            // 対象のtodoにedit後のtodoを反映
-            targetTodoArray
-              .filter((todo) => todo.id == todoEdited.id)
-              .forEach((todo) => {
-                todo.title = todoEdited.title;
-                todo.content = todoEdited.content;
-                todo.category = todoEdited.category;
-                todo.date_limit = todoEdited.date_limit;
-              });
-
-            return {
-              ...prevTodos,
-              [todoEdited.category]: targetTodoArray,
-            };
-          } else {
-            // edit前と後でcategoryが違う場合
-            // 'targetTodoArray'より対象のtodoのインデックスの取得
-            const targetTodoIndex = prevTodos[prevTodoCategory].findIndex(
-              (prevTodo) => prevTodo.id == todoEdited.id
-            );
-            // 対象のtodoの取得
-            let targetTodo = prevTodos[prevTodoCategory].splice(
-              targetTodoIndex,
-              1
-            );
-            targetTodo.forEach((todo) => {
-              todo.title = todoEdited.title;
-              todo.content = todoEdited.content;
-              todo.category = todoEdited.category;
-              todo.date_limit = todoEdited.date_limit;
-            });
-
-            return {
-              ...prevTodos,
-              [todoEdited.category]:
-                prevTodos[todoEdited.category].concat(targetTodo),
-            };
-          }
-        });
-
-        // フォーム画面を閉じる
-        setIsShowForm(false);
-      } else {
-        // AxiosError
-        if (data.validationErrors) {
-          // バリデーションエラー
-          const validationErrors = data.validationErrors;
-          handleValidationErrors(validationErrors!);
-        } else {
-          alert(`ERROR: ${putTodoResult.status}`);
-        }
-      }
-    } else {
-      // Error
-      alert("ERROR");
-    }
-  };
-
-  // バリデーションエラーの内容をstate'errorMessages'に反映する
-  const handleValidationErrors = (
-    responseValidationErrors: ValidationErrors
-  ) => {
-    let validationErrorMessages: ErrorMessages = {
-      title: [],
-      content: [],
-      date_limit: [],
-    };
-    // validationErrorsのメッセージをvalidationErrorMessagesに追加していく
-    Object.keys(responseValidationErrors).forEach((key) => {
-      const validationErrors = responseValidationErrors[key];
-      validationErrors.forEach((validationError: ValidationError) => {
-        validationErrorMessages[key].push(validationError.message);
-      });
-    });
-    // validationErrorMessagesをerrorMessagesにセットする
-    setErrorMessages(validationErrorMessages);
-  };
-
-  // todoの削除。削除後、画面に反映
-  const deleteTodo: FnToHandleTodosTable = async () => {
-    const deleteTodoResult = await TodoApi.deleteTodo(todo);
-
-    if (deleteTodoResult) {
-      const data = deleteTodoResult.data;
-      if (data.status == "success") {
-        const todoDeleted = data.todoDeleted;
-        setTodos((prevTodos) => {
-          // 削除されたtodoが格納されている配列
-          let targetTodoArray = prevTodos[prevTodoCategory];
-          // 削除されたtodoのインデックスの取得
-          const targetTodoIndex = targetTodoArray.findIndex(
-            (todo) => todo.id == todoDeleted.id
-          );
-          // 削除
-          prevTodos[prevTodoCategory].splice(targetTodoIndex, 1);
-
-          return {
-            ...prevTodos,
-            [prevTodoCategory]: prevTodos[prevTodoCategory],
-          };
-        });
-      } else {
-        // AxiosError
-        alert(`ERROR: ${deleteTodoResult.status}`);
-      }
-    } else {
-      // Error
-      alert("ERROR");
-    }
-  };
-
   return (
     <div css={todosContainer}>
       {isShowForm && (
         <TodosForm
-          todo={todo}
-          setTodo={setTodo}
-          setIsShowForm={setIsShowForm}
-          todoFunction={formType == "new" ? createTodo : editTodo}
+          formType={formType}
+          prevTodoCategory={prevTodoCategory}
           errorMessages={errorMessages}
+          setIsShowForm={setIsShowForm}
           setErrorMessages={setErrorMessages}
         />
       )}
       {isShowTodoDesc && (
         <TodoDescription
-          todo={todo}
           setIsShowTodoDesc={setIsShowTodoDesc}
           setIsShowForm={setIsShowForm}
           setFormType={setFormType}
-          deleteTodo={deleteTodo}
+          prevTodoCategory={prevTodoCategory}
         />
       )}
       {/* DndProviderでラップされいるコンポーネント間でdrag&dropが可能 */}
@@ -343,9 +152,7 @@ const TodosContens: FC = () => {
           currentPage={currentPageShort}
           maxPage={maxPageShort}
           setPage={setPageShort}
-          setTodo={setTodo}
           setFormType={setFormType}
-          setTodos={setTodos}
         >
           {returnTodosForColumn(TODO_CATEGORIES.SHORT, currentPageShort)}
         </TodosCategoryColumn>
@@ -355,9 +162,7 @@ const TodosContens: FC = () => {
           currentPage={currentPageMedium}
           maxPage={maxPageMedium}
           setPage={setPageMedium}
-          setTodo={setTodo}
           setFormType={setFormType}
-          setTodos={setTodos}
         >
           {returnTodosForColumn(TODO_CATEGORIES.MEDIUM, currentPageMedium)}
         </TodosCategoryColumn>
@@ -367,9 +172,7 @@ const TodosContens: FC = () => {
           currentPage={currentPageLong}
           maxPage={maxPageLong}
           setPage={setPageLong}
-          setTodo={setTodo}
           setFormType={setFormType}
-          setTodos={setTodos}
         >
           {returnTodosForColumn(TODO_CATEGORIES.LONG, currentPageLong)}
         </TodosCategoryColumn>
@@ -379,15 +182,13 @@ const TodosContens: FC = () => {
           currentPage={currentPageCompleted}
           maxPage={maxPageCompleted}
           setPage={setPageCompleted}
-          setTodo={setTodo}
           setFormType={setFormType}
-          setTodos={setTodos}
         >
           {returnTodosForColumn(TODO_CATEGORIES.COMPLETE, currentPageCompleted)}
         </TodosCategoryColumn>
       </DndProvider>
     </div>
   );
-};
+}
 
-export default TodosContens;
+export default TodosContents;
