@@ -1,13 +1,14 @@
-import { FC, Dispatch, SetStateAction } from "react";
+import { Dispatch, SetStateAction } from "react";
 import { css } from "@emotion/react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faRectangleXmark } from "@fortawesome/free-solid-svg-icons/faRectangleXmark";
 import { faPenToSquare } from "@fortawesome/free-solid-svg-icons/faPenToSquare";
 import { faTrash } from "@fortawesome/free-solid-svg-icons/faTrash";
 import Tooltip from "../../components/layout/Tooltip/Tooltip";
-import { Todo } from "./TodosContents";
+import { useTodo } from "../../components/context/TodoContext";
 import { TODO_CATEGORIES } from "../../utils/constants/todoCategory.constants";
 import { DateFormatters } from "../../utils/helpers/date.helpers";
+import { TodoApi } from "../../services/api/todoApi";
 
 const todoDescContainer = css({
   height: "100%",
@@ -87,13 +88,20 @@ const closeBtn = css({
   },
 });
 
-const TodoDescription: FC<{
-  todo: Todo;
-  setIsShowTodoDesc: Dispatch<SetStateAction<boolean>>;
-  setIsShowForm: Dispatch<SetStateAction<boolean>>;
-  setFormType: Dispatch<SetStateAction<string>>;
-  deleteTodo: FnToHandleTodosTable;
-}> = ({ todo, setIsShowTodoDesc, setIsShowForm, setFormType, deleteTodo }) => {
+function TodoDescription(
+  {
+    setIsShowTodoDesc,
+    setIsShowForm,
+    setFormType,
+    prevTodoCategory,
+  }: {
+    setIsShowTodoDesc: Dispatch<SetStateAction<boolean>>;
+    setIsShowForm: Dispatch<SetStateAction<boolean>>;
+    setFormType: Dispatch<SetStateAction<string>>;
+    prevTodoCategory: string;
+  }) {
+  const { todo, setTodosByCategory } = useTodo();
+
   const handleEditIcon = () => {
     // todo詳細ページを非表示
     setIsShowTodoDesc(false);
@@ -108,6 +116,39 @@ const TodoDescription: FC<{
     deleteTodo();
     // todo詳細ページを非表示
     setIsShowTodoDesc(false);
+  };
+
+  // todoの削除。削除後、画面に反映
+  const deleteTodo = async () => {
+    const deleteTodoResult = await TodoApi.deleteTodo(todo);
+
+    if (deleteTodoResult) {
+      const data = deleteTodoResult.data;
+      if (data.status == "success") {
+        const todoDeleted = data.todoDeleted;
+        setTodosByCategory((prevTodos) => {
+          // 削除されたtodoが格納されている配列
+          const targetTodoArray = prevTodos[prevTodoCategory];
+          // 削除されたtodoのインデックスの取得
+          const targetTodoIndex = targetTodoArray.findIndex(
+            (todo) => todo.id == todoDeleted.id
+          );
+          // 削除
+          prevTodos[prevTodoCategory].splice(targetTodoIndex, 1);
+
+          return {
+            ...prevTodos,
+            [prevTodoCategory]: prevTodos[prevTodoCategory],
+          };
+        });
+      } else {
+        // AxiosError
+        alert(`ERROR: ${deleteTodoResult.status}`);
+      }
+    } else {
+      // Error
+      alert("ERROR");
+    }
   };
 
   return (
@@ -148,11 +189,11 @@ const TodoDescription: FC<{
         </div>
         <div css={todoDescContent}>
           <label>Deadline</label>
-          <p>{DateFormatters.convertStrDateToDispDate(todo.date_limit)}</p>
+          <p>{DateFormatters.convertISOStringToDispDate(todo.date_limit)}</p>
         </div>
         <div css={todoDescContent}>
-          <label>Date Created</label>
-          <p>{DateFormatters.convertUtcStrDateToDispDate(todo.date_created)}</p>
+          <label>Created at</label>
+          <p>{DateFormatters.convertISOStringToDispDate(todo.created_at)}</p>
         </div>
         <div css={closeBtnkWrapper}>
           <FontAwesomeIcon
@@ -165,6 +206,6 @@ const TodoDescription: FC<{
       </div>
     </div>
   );
-};
+}
 
 export default TodoDescription;
